+++ backend/lib/edgehog/geolocation/providers/device_geolocation.ex
@@ -31,9 +31,8 @@ defmodule Edgehog.Geolocation.Providers.DeviceGeolocation do
   def geolocate(%Device{device_id: device_id} = device) do
     with {:ok, client} <- Devices.appengine_client_from_device(device),
          {:ok, sensors_positions} <- Astarte.fetch_geolocation(client, device_id),
-         {:ok, sensors_positions} <- filter_latest_sensors_positions(sensors_positions),
-         {:ok, position} <- geolocate_sensors(sensors_positions) do
-      {:ok, position}
+         {:ok, sensors_positions} <- filter_latest_sensors_positions(sensors_positions) do
+      geolocate_sensors(sensors_positions)
     end
   end