+++ backend/lib/edgehog/geolocation/providers/google_geolocation.ex
@@ -37,9 +37,8 @@ defmodule Edgehog.Geolocation.Providers.GoogleGeolocation do
   def geolocate(%Device{device_id: device_id} = device) do
     with {:ok, client} <- Devices.appengine_client_from_device(device),
          {:ok, wifi_scan_results} <- Astarte.fetch_wifi_scan_results(client, device_id),
-         {:ok, wifi_scan_results} <- filter_latest_wifi_scan_results(wifi_scan_results),
-         {:ok, position} <- geolocate_wifi(wifi_scan_results) do
-      {:ok, position}
+         {:ok, wifi_scan_results} <- filter_latest_wifi_scan_results(wifi_scan_results) do
+      geolocate_wifi(wifi_scan_results)
     end
   end