+++ backend/lib/edgehog/geolocation/providers/ipbase.ex
@@ -60,9 +60,8 @@ defmodule Edgehog.Geolocation.Providers.IPBase do
 
   defp geolocate_ip(ip_address) do
     with {:ok, api_key} <- Config.ipbase_api_key(),
-         {:ok, %{body: body}} <- get("", query: [apikey: api_key, ip: ip_address]),
-         {:ok, location} <- parse_response_body(body) do
-      {:ok, location}
+         {:ok, %{body: body}} <- get("", query: [apikey: api_key, ip: ip_address]) do
+      parse_response_body(body)
     end
   end
 
@@ -78,8 +77,7 @@ defmodule Edgehog.Geolocation.Providers.IPBase do
 
       address =
         [city, zip, region, country]
-        |> Enum.reject(&is_nil/1)
-        |> Enum.reject(&(&1 == ""))
+        |> Enum.reject(&(is_nil(&1) or &1 == ""))
         |> Enum.join(", ")
 
       location = %{