+++ backend/test/edgehog/astarte/device/storage_usage_test.exs
@@ -21,9 +21,9 @@
 defmodule Edgehog.Astarte.Device.StorageUsageTest do
   use Edgehog.DataCase
 
+  alias Astarte.Client.AppEngine
   alias Edgehog.Astarte.Device.StorageUsage
   alias Edgehog.Astarte.Device.StorageUsage.StorageUnit
-  alias Astarte.Client.AppEngine
 
   describe "storage_usage" do
     import Edgehog.AstarteFixtures
@@ -64,7 +64,7 @@ defmodule Edgehog.Astarte.Device.StorageUsageTest do
       assert {:ok, storage_units} = StorageUsage.get(appengine_client, device.device_id)
 
       assert storage_units == [
-               %StorageUnit{free_bytes: 7000, label: "nvs1", total_bytes: 16128}
+               %StorageUnit{free_bytes: 7000, label: "nvs1", total_bytes: 16_128}
              ]
     end
 
@@ -95,7 +95,7 @@ defmodule Edgehog.Astarte.Device.StorageUsageTest do
       assert {:ok, storage_units} = StorageUsage.get(appengine_client, device.device_id)
 
       assert storage_units == [
-               %StorageUnit{free_bytes: 7000, label: "nvs1", total_bytes: 16128},
+               %StorageUnit{free_bytes: 7000, label: "nvs1", total_bytes: 16_128},
                %StorageUnit{free_bytes: 5000, label: "nvs2", total_bytes: 8064}
              ]
     end