+++ backend/test/edgehog/devices_test.exs
@@ -644,8 +644,8 @@ defmodule Edgehog.DevicesTest do
 
   describe "system_models" do
     alias Edgehog.Devices.SystemModel
-    alias Edgehog.Devices.SystemModelPartNumber
     alias Edgehog.Devices.SystemModelDescription
+    alias Edgehog.Devices.SystemModelPartNumber
 
     import Edgehog.AstarteFixtures, except: [device_fixture: 1, device_fixture: 2]
     import Edgehog.DevicesFixtures
@@ -844,7 +844,7 @@ defmodule Edgehog.DevicesTest do
 
       assert [system_model_1, system_model_2] =
                Devices.list_system_models()
-               |> Devices.preload_localized_descriptions_for_system_model("it-IT")
+               |> Devices.preload_localized_descriptions_for_system_model(["it-IT"])
 
       assert [%{locale: "it-IT"}] = system_model_1.descriptions
       assert [] = system_model_2.descriptions
@@ -864,9 +864,12 @@ defmodule Edgehog.DevicesTest do
       assert {:ok, system_model} = Devices.fetch_system_model(system_model.id)
 
       system_model =
-        Devices.preload_localized_descriptions_for_system_model(system_model, "it-IT")
+        Devices.preload_localized_descriptions_for_system_model(system_model, ["it-IT", "en-US"])
 
-      assert Enum.map(system_model.descriptions, & &1.locale) == ["it-IT"]
+      assert [
+               %{locale: "en-US", text: "A system model"},
+               %{locale: "it-IT", text: "Un modello di sistema"}
+             ] = system_model.descriptions
     end
   end
 end